Исправила все замечания
