+++ src/test/java/UpdateAccountTest.java
@@ -1,42 +1,40 @@
 import io.qameta.allure.junit4.DisplayName;
 import io.restassured.response.ValidatableResponse;
-import org.example.metods.Login;
-import org.example.metods.СhangeAccount;
+import org.example.metods.CreateAccount;
 import org.example.model.Credentials;
-import org.example.model.Data;
+import org.example.Data;
 import org.junit.Before;
 import org.junit.Test;
 import static org.hamcrest.Matchers.equalTo;
 import static org.hamcrest.Matchers.is;
 import static org.junit.Assert.assertEquals;
+import static org.apache.http.HttpStatus.*;
 
 public class UpdateAccountTest {
-    private Login login;
     private String token;
-    private СhangeAccount сhangeAccount;
     private Credentials credentials;
+    private CreateAccount createAccount;
 
     @Before
     public void setUp() {
         credentials = new Credentials(Data.LOGIN, Data.PASSWORD);
-        login = new Login();
-        сhangeAccount = new СhangeAccount();
+        createAccount = new CreateAccount();
     }
     @Test
     @DisplayName("клиент может изменить данные с авторизацией")
     public void updateAccount() {
-        ValidatableResponse response = login.login(credentials);
+        ValidatableResponse response = createAccount.login(credentials);
         token = response.extract().path("accessToken");
-        ValidatableResponse response1 = сhangeAccount.change(token, new Credentials(Data.LOGIN, Data.PASSWORD));
+        ValidatableResponse response1 = createAccount.change(token, new Credentials(Data.LOGIN, Data.PASSWORD));
         response1.assertThat().body("success", is(true));
     }
     @Test
     @DisplayName("клиент не может изменить данные без авторизацией")
     public void NotUpdateAccount() {
 
-        ValidatableResponse response1 = сhangeAccount.change("", new Credentials(Data.LOGIN, Data.PASSWORD));
+        ValidatableResponse response1 = createAccount.change("", new Credentials(Data.LOGIN, Data.PASSWORD));
         int statusCode = response1.extract().statusCode();
         response1.assertThat().body("success", is(false)).and().body("message", equalTo("You should be authorised"));
-        assertEquals("Fail", 401, statusCode);
+        assertEquals("Fail", SC_UNAUTHORIZED, statusCode);
     }
 }
